# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_checkpoint_server_internal_models.Activity:
    properties:
      action_type:
        description: rated, reviewed, status_change
        type: string
      created_at:
        type: string
      id:
        type: string
      target_id:
        type: string
      target_type:
        description: review, user_game
        type: string
      user:
        $ref: '#/definitions/github_com_checkpoint_server_internal_models.User'
      user_id:
        type: string
    type: object
...
  github_com_checkpoint_server_internal_models.User:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      steam_id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: The Game Journal (Letterboxd for games) API
  title: Checkpoint MVP API
  version: "1.0"
paths:
...
  /social/activity/{user_id}:
    get:
      consumes:
      - application/json
      description: Get the activity feed for a specific user
      parameters:
      - description: User UUID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_checkpoint_server_internal_models.Activity'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get User Activity
      tags:
      - social
  /social/feed:
    get:
      consumes:
      - application/json
      description: Get activity feed of followed users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_checkpoint_server_internal_models.Activity'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Activity Feed
      tags:
      - social
  /social/follow:
    post:
      consumes:
      - application/json
      description: Follow another user
      parameters:
      - description: User UUID to follow
        in: query
        name: following_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Follow User
      tags:
      - social
  /social/profile/{user_id}:
    get:
      consumes:
      - application/json
      description: Fetch public user profile and basic stats
      parameters:
      - description: User UUID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get User Profile
      tags:
      - social
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"